requirements.txt
asttokens==3.0.0
colorama==0.4.6
executing==2.1.0
fonttools==4.55.3
icecream==2.1.3
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
Pygments==2.18.0
pytesseract==0.3.13
watchdog==6.0.0
